reinit the app
