--- README.md
@@ -1,1 +1,0 @@
-# CS5520-Ast2
