+++ app.json
@@ -1,7 +1,7 @@
 {
   "expo": {
-    "name": "ast2-calorie-app",
-    "slug": "ast2-calorie-app",
+    "name": "ast2",
+    "slug": "ast2",
     "version": "1.0.0",
     "orientation": "portrait",
     "icon": "./assets/icon.png",
