--- components/Button.js
@@ -1,25 +1,0 @@
-import React from 'react';
-import { TouchableOpacity, Text, StyleSheet } from 'react-native';
-
-const Button = ({ onPress, title, style, textStyle }) => (
-  <TouchableOpacity onPress={onPress} style={[styles.button, style]}>
-    <Text style={[styles.buttonText, textStyle]}>{title}</Text>
-  </TouchableOpacity>
-);
-
-const styles = StyleSheet.create({
-  button: {
-    backgroundColor: '#4caf50',
-    paddingVertical: 10,
-    paddingHorizontal: 20,
-    borderRadius: 5,
-  },
-  buttonText: {
-    color: 'white',
-    fontWeight: 'bold',
-    textTransform: 'uppercase',
-    textAlign: 'center',
-  },
-});
-
-export default Button;
