+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "ast2-calorie-app",
+  "name": "ast2",
   "version": "1.0.0",
   "main": "node_modules/expo/AppEntry.js",
   "scripts": {
@@ -9,17 +9,10 @@
     "web": "expo start --web"
   },
   "dependencies": {
-    "@react-navigation/bottom-tabs": "^6.5.4",
-    "@react-navigation/native": "^6.1.3",
-    "@react-navigation/native-stack": "^6.9.9",
-    "@react-navigation/stack": "^6.3.12",
     "expo": "~47.0.12",
     "expo-status-bar": "~1.4.2",
-    "firebase": "^9.17.1",
     "react": "18.1.0",
-    "react-native": "0.70.5",
-    "react-native-safe-area-context": "4.4.1",
-    "react-native-screens": "~3.18.0"
+    "react-native": "0.70.5"
   },
   "devDependencies": {
     "@babel/core": "^7.12.9"
